wlcore: add version and timestamp strings (INTERNAL)
In order to allow better version control, add version
and timestamp strings into wlcore module.

These strings will be printed on driver load, and when
reading the driver_state.

If the tree was changed and unstaged or if a new tag was added, we
were not regenerating the version.h file, so it would contain outdated
information.

To fix this, force the version.h file to be rechecked during make and
use __TIMESTAMP__ instead of using the shell's date command, so that
we don't regenerate it everytime.

Additionally, use --dirty in git describe to mark dirty trees as such.

With compat-wireless, we shouldn't force the version.h file checks,
otherwise it will overwrite what we generated in compat-wireless
itself.  This was only needed for dirty trees anyway.  The side-effect
is that the version will not be regenerated when the tree becomes
dirty.

In environments without .git folder, use release_version.h instead.

There are some reports of compilation errors due
to missing version.h.
This probably means that version.h wasn't recognized
as a dependency, and wasn't generated.
Declare this dependency explicitly in order to make
sure version.h will be generated.

if we use backports to compile our driver, the compilation
is done on the target kernel, but we want to generate the
version tag from the driver git.

check for the existence of KLIB_BUILD and GIT_TREE params,
and generate version.h from the GIT_TREE if they both
exist.
(note that backports doesn't require GIT_TREE to be
defined, so it has to be part of the build script)

When checking for git existence in wlcore's Makefile,
use $(PWD), as the built-in wildcard function returns
unexpected results when working against remote
makefile (e.g. when using backports).

git describe --dirty doesn't work well when specfiying
git-dir and work-tree parameters (git diff-index doesn't
run against the work-tree, resulting in -dirty always
being set).

workaround it by changing the current directory to the
git directory before running the command.

Signed-off-by: Eliad Peller <eliad@wizery.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>
Signed-off-by: Arik Nemtsov <arik@wizery.com>