wlcore: use legacy mode for all ACs
AVM FritzBox 7390 has a weird behaviour where once
we enable U-APSD only on the VO AC it limits the rate
of the packets it sends to 39Mbps and doesn't do AMPDU
aggregation. This causes a big degradation in throughput
obviously.

It's important to note that this isn't related to PS
and happens also when ACTIVE is forced. The traffic
affected is sent in BE AC. Once legacy mode for all ACs
is used the AP begins using higher rates as well as AMPDU
aggregation which improves throughput significantly.

In order to deal with this AP and potentially others
switch to using legacy mode for all ACs by default.

Signed-off-by: Eyal Shapira <eyal@wizery.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>