wlcore: CAC only in case of usable DFS channel
The current flow of configuring the ap (including channel
contexts) after the initial CAC is completed, results in
another quick CAC (start + stop) being sent, which
the phy doesn't handle well.

Look for the channel dfs state when radar_enabled is set,
and start CAC only if the channel is usable.

Signed-off-by: Eliad Peller <eliad@wizery.com>