wl12xx : move platform_data->gpio to match old kernels (INTERNAL)
Older kernels don't have the gpio field defined.
In order to be able to work against them, move the gpio
field to the end of the struct, and add a comment
to avoid using it if unsure.

(it is safe to use it if the whole struct is defined
by the driver, specifically in the case of device-tree
without platform_data defined in the kernel)

Signed-off-by: Eliad Peller <eliad@wizery.com>